requirements.txt
# # Core server and API framework
# fastapi
# uvicorn

# # OpenRouter-compatible LLM client
# openai>=1.0.0

# # Configuration and schema validation
# pydantic
# python-dotenv

# # Long-term memory via vector store
# chromadb

# # Async HTTP requests (for hosted tools or APIs)
# httpx

# # Optional: development and formatting tools
# black
# isort
# pytest
# ipython
# typer

# # Optional: embeddings and vector math (for RAG or semantic memory)
# sentence-transformers
# numpy

# # web tool
# httpx>=0.27.0
# beautifulsoup4>=4.12.3


# Core typing and validation for tools and models
pydantic>=2.6.0,<3

# Async HTTP client for web tools (search/scraper)
httpx>=0.27.0,<0.28

# HTML parsing for the scraper and search result parsing
beautifulsoup4>=4.12.3,<5

# OpenAI-compatible client used with OpenRouter (base_url=https://openrouter.ai/api/v1)
openai>=1.40.0,<2
# YAML parsing for configuration files
PyYAML>=6.0.1,<7

python-dotenv>=1.0.1,<2




